Separation of concerns
Refactored ForecastToday and ForecastList to only accept one `item`,
and have the App pass them in. ForecastList can be renamed to
ForecastListItem but I wanted to reduce the changeset so it wasn't
confusing.

This change makes the components "dumber" and App can also assign
the keys itself when mapping over the week.

Unable to test due to the API key not being on my machine.